+++ examples/typescript/core/src/controllers/posts/index.ts
@@ -1,0 +1,6 @@
+import App from '../../app';
+import { collection } from './state';
+
+export default App.Controller({
+  collection
+});