+++ packages/pulse-core/lib/event.ts
@@ -1,7 +1,7 @@
 import { Pulse } from './internal';
 
 // default event payload
-export type EventPayload = { [key: string]: any };
+export type EventPayload = any;
 // type of the callback an event should receive
 export type EventCallbackFunc<P = EventPayload> = (payload: P) => void;
 // type of the function used as an alias to create an event instance, with a generic for the payload type
@@ -30,6 +30,8 @@ export class Event<P = EventPayload> {
   private queue: Array<P>;
   // should never be defined, but holds reference to the Payload type for useEvent to read
   public payload: P;
+
+  private onNextCallback: (payload?: P) => any;
 
   constructor(public instance: () => Pulse, public config: EventConfig<P> = {}) {
     // initiate uses state if applicable
@@ -72,10 +74,19 @@ export class Event<P = EventPayload> {
     this.config.enabled = false;
   }
 
+  public onNext(callback: (payload?: P) => any) {
+    this.onNextCallback = callback;
+  }
+
   // Private functions
   private emitter(payload: P) {
     // foreach callback, invoke the saved function
     this.callbacks.forEach(callback => callback(payload));
+
+    if (typeof this.onNextCallback === 'function') {
+      this.onNextCallback(payload);
+      delete this.onNextCallback;
+    }
     // increment the uses if
     if (this.uses !== undefined) this.uses++;
   }