+++ packages/pulse-core/lib/pulse.ts
@@ -1,5 +1,7 @@
 // prettier-ignore
-import { State, StateGroup, Computed, Collection, GroupObj, DefaultDataItem, SelectorObj, Config, SubController, Runtime, Storage, Event, EventPayload, EventConfig, EventsObjFunc, StorageConfig, API, APIConfig, Group, Controller, ControllerConfig, FuncObj, StateObj, StatusTracker, Integration, Integrations } from './internal';
+import { State, StateGroup, Computed, Collection, GroupObj, DefaultDataItem, SelectorObj, Config, SubController, Runtime, Storage, Event, EventPayload, EventConfig, EventsObjFunc, StorageConfig, API, APIConfig, Group, Controller, ControllerConfig, FuncObj, StateObj, StatusTracker, Integration, Integrations, Action, FuncType } from './internal';
+import { Tracker } from './tracker';
+import { HistoryItem } from './state';
 
 export interface PulseConfig {
   computedDefault?: any;
@@ -7,6 +9,7 @@ export interface PulseConfig {
   storage?: StorageConfig;
   logJobs?: boolean;
   noCore?: boolean;
+  globalHistory?: boolean;
 }
 
 export const defaultConfig: PulseConfig = {
@@ -28,6 +31,8 @@ export class Pulse {
   public controllers: { [key: string]: any } = {};
   public subController: SubController;
   public errorHandlers: Set<(error: ErrorObject) => void> = new Set();
+
+  public history: HistoryItem[] = [];
 
   // integrations
   public integrations: Integrations;
@@ -77,7 +82,7 @@ export class Pulse {
    * @param deps Array - An array of state items to depend on
    * @param func Function - A function where the return value is the state, ran every time a dep changes
    */
-  public Computed<T = any>(func: () => any, deps?: Array<any>) {
+  public Computed<T = any>(func: () => T, deps?: Array<any>) {
     const computed = new Computed<T>(() => this, func, deps);
     this._computed.add(computed);
     return computed;
@@ -99,12 +104,8 @@ export class Pulse {
   /**
    * Create a Pulse Action
    */
-  public Action(func: Function) {
-    return () => {
-      const returnValue = func();
-
-      return returnValue;
-    };
+  public Action<T extends FuncType>(func: T) {
+    return new Action(() => new Pulse(), func).hoc();
   }
 
   /**
@@ -152,7 +153,7 @@ export class Pulse {
   /**
    * Create a Pulse Error
    */
-  public Error(error: any, code?: string) {}
+  public Error(error: any, info?: { fromAction?: any }) {}
 
   /**
    * onError handler
@@ -190,6 +191,13 @@ export class Pulse {
     this.storage = new Storage(() => this, config);
     this.storage.persistedState = persistedState;
     this.storage.persistedState.forEach(state => state.persist(state.name));
+  }
+
+  public track(changeFunc: () => void) {
+    return new Tracker(() => this, changeFunc);
+  }
+  public batch(batchFunc: () => void): void {
+    this.runtime.batch(batchFunc);
   }
 
   /**