+++ docs/v3/introduction/changelog.md
@@ -1,19 +1,35 @@
 ---
 title: Changelog
 ---
-
-## 3.4 - The one with VueJS support
-
+## 3.5 - The one with Actions
 ::: warning Work in progress
 This update is currently being worked on, but will be released soon!
 :::
 
-- Added support for VueJS
-- Added State.record() to optionally track mutation history
-- Added `useWatcher()` for React integration
-- Watchers can now accept callbacks as the first parameter, returning a unique key for cleanup (non breaking)
+**NEW FEATURES**
+- Added [`App.Action()`]() - A better way to write actions. [still WIP]
+- Added [`App.Error()`]() - A global error handler. [still WIP]
+- Added [`App.track()`]() - Track multiple mutations and revert using returned [`undo()`]() method.
+- Added [`App.batch()`]() - Increase performance by batching multiple mutations.*
+- Added [`Collection.onCollect()`]() - Mutate data on collect
+> onCollect is a persistent alternative to [`Collection.compute()`]() which runs the mutate function only on Group output, not affecting original data. onCollect will mutate each data item collected as it is collected.
+- Added [`Event.onNext()`]() - Run a callback once after next event.
 
-## 3.3 - The one with [@pulsejs]() imports
+**BIG PERFORMANCE GAINS**: 
+- Groups can now _**soft rebuild**_ when generating output, avoiding rebuilding from scratch every time data was changed. Group will memorize the change key & index, applying a custom splice action to the cached output data.
+- *`App.batch()` Can be used to instruct runtime to batch State mutations together, this will result in the side effects being squashed and offset to the end of the batch.
+- Removed a `Collection.getGroup()` memory leak when used inside a Computed function, where no group is found. It would return an empty group that is dependent on the Computed function, which would add a useless dependent every recompute. Now `getGroup()` memorizes the requested key by creating a provisional group registered in the Collection.
+
+**MISC FIXES**:
+- [Computed]() will now generate immediately if created _after_ core has initialized (App.Core())
+- [`Group.index`]() now returns the index accurate to the output.
+    > Useful if missing data was encountered during Group build. Previously it was an alias for `Group.value` directly, which could contain primary keys for data that does not exist in the collection. `Group.index` is now always true to `Group.output`, while `Group.value` contains the primary keys desired to be in a group, even if they don't exist in the Collection.
+- Fixed bug with API class that was causing one-time header overrides to persist to all future route calls.
+- API class will allow you to set content-type, previously it forced auto detection.
+- [`State.interval()`]() now returns _this_, saves interval id locally and provides [`State.clearInterval()`]().
+- Collection's default group will not be auto created unless specified in config OR no groups are provided at initialization.
+
+## 3.4 - The one with useWatcher()
 
 ::: tip Current NPM version
 This is the current release build available on NPM
@@ -24,6 +40,15 @@ yarn add @pulsejs/core
 
 See updated integration with [React](../getting-started/setup-with-react.html) and [Next](../getting-started/setup-with-next.html)
 :::
+
+- Added State.record() to optionally track mutation history
+- Added `useWatcher()` for React integration
+- Watchers can now accept callbacks as the first parameter, returning a unique key for cleanup (non breaking)
+- Collect method now preserves the correct order when collecting more than one item in unshift mode
+- Collection `forEachItem` function now allows you to return a modified object
+- Computed now throws and error if no function is provided
+
+## 3.3 - The one with [@pulsejs]() imports
 
 - Deprecated NPM package `pulse-framework`
 - Converted Pulse to monorepo with new NPM org ([@pulsejs](https://www.npmjs.com/org/pulsejs))